メモリー資源の見直しと超適当にノイズを追加
エンベロープなど非実装。今後追加します。
ノイズは波形にしては周期が非常に短すぎて使えない。。（ブーって音が出てしまう)

+++ audio.h
@@ -45,9 +45,9 @@ typedef struct{
 
 typedef struct{
   unsigned int time;
-  unsigned int len;
-  unsigned int key;
-  unsigned int vel;
+  uint16_t len;
+  uint8_t key;
+  uint8_t vel;
 }music_dat_t;
